Allow multiple connections to file databases
Use fbcdcConnectToURL() when connecting to a file database as that functions
allows a single process to open several concurrent connections to the same
database.

Also ensure that the error messages returned from fbsConnectDatabaseAtPath
and fbsExecuteSQL are proper copies of the FB error messages, so that
callers can dispose of them with a normal call to `free()`.

This last point applies to other functions returning error messages too
and will be addressed in a future commit.